examples/python/requirements.txt
# pip install -r requirements.txt

requests~=2.25.1
websocket-client~=1.3.3